+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 # Release History
 
 ## v1.4.0
-> Release Date: Future
+> Release Date: Unreleased
 
 Enhancement:
-- Added support for VMware Cloud Foundation 5.0 on VxRail.
+- Added support for VMware Cloud Foundation 5.0 on Dell VxRail.
 
 ## v1.3.0